Create a directory

Create a directory if it does not yet exist.

Usage

create_dir(dir = fs::path(".", "output"), add_date = TRUE)

Arguments

dir

Character string containing a valid path to a directory that should be created if it does not yet exist. fs::path() adds file separators and the dot (".") indicates the working directory, such that by default a subdirectory with the current date in the format YYYY_mm_dd in directory output below the working directory is created.

add_date

TRUE or FALSE: create a subdirectory in dir with the current date in the format YYYY_mm_dd?

Value

A character string with the absolute normalized path to the requested directory, returned invisibly. An error is thrown if the attempt to create a directory fails, for example because dir points to an existing file instead of to an directory.

Details

The absolute, normalised path is returned such that the returned path still works if the working directory changes. On case-insensitive file systems, normalization also adjusts the case to match names of directories that are already present (see the Examples).

Side effects

The directory indicated by the returned path is created if it does not yet exist.

See also

Examples

# Use a temporary subdirectory to not write in the user's directory
my_tempdir <- create_tempdir(prefix = "examplecreatedir")

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                          add_date = FALSE)
fs::dir_exists(res_dir_one) # TRUE
#> /tmp/RtmpoivS2Y/examplecreatedir1a0e4f81398d/dir_one 
#>                                                 TRUE 

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned.
res_dir_one_v2 <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE
#> [1] TRUE

# On case-insensitive file systems, the directory 'res_dir_ONE' is the same as
# 'res_dir_one'. On case-sensitive file systems it differs in case from
# 'res_dir_one'.
res_dir_ONE <- create_dir(dir = fs::path(my_tempdir, "dir_ONE"),
                          add_date = FALSE)
identical(res_dir_one, res_dir_ONE)
#> [1] FALSE

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_date <- create_dir(dir = fs::path(my_tempdir, "dir_date"),
                           add_date = TRUE)
fs::dir_exists(res_dir_date) # TRUE
#> /tmp/RtmpoivS2Y/examplecreatedir1a0e4f81398d/dir_date/2026_06_24 
#>                                                             TRUE 

# Cleaning up
unlink(my_tempdir, recursive = TRUE)
rm(my_tempdir, res_dir_one, res_dir_one_v2, res_dir_date, res_dir_ONE)