Remove extension from file paths

A drop-in replacement for tools::file_path_sans_ext, see Details. Returns the file path without the extensions (and the leading dot). Only purely alphanumeric extensions are recognized.

Usage

file_path_sans_ext(x, compression = FALSE)

Arguments

x

character vector giving file paths.

compression

logical: should compression extension .gz, .bz2 or .xz be removed first?

Details

tools::file_path_sans_ext() did not recognise the extension of file names that end in a dot prior to R 4.6.0, whereas tools::file_ext() did recognise such extensions.

Using file_path_sans_ext() from progutils ensures that filename is recreated by paste0(progutils::file_path_sans_ext(filename), ".", tools::file_ext(filename)), such that, e.g., create_path("ab..txt") produces the correct result ending in "ab..txt" instead of the nonsense result ending in "ab..txt.txt", see the Examples.

See also

Examples

filename <- "ab..txt"
# Should be "ab." but was "ab..txt" prior to R 4.6.0.
tools::file_path_sans_ext(filename)
#> [1] "ab."
tools::file_ext(filename) # "txt"
#> [1] "txt"
# So the next line produces the nonsense-result "ab..txt.txt"
paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))
#> [1] "ab..txt"

# The updated version recreates filename 'ab..txt':
paste0(progutils::file_path_sans_ext(filename), ".", tools::file_ext(filename))
#> [1] "ab..txt"