Check that only a single matching file is present

Check that only one file in a directory has a name matching pattern, for example before attempting to read a file.

Usage

get_file_path(dir = ".", pattern, ignore_case = TRUE, quietly = FALSE)

Arguments

dir

Character string with the path to a directory.

pattern

Character string containing a regular expression used to select names of files that are present in dir.

ignore_case

TRUE or FALSE: use case-insensitive pattern matching?

quietly

TRUE or FALSE: suppress the message with the found file name?

Value

A character string with the normalized path to the file with a name matching pattern if there is exactly one such file in directory dir. Otherwise an error is thrown. Use basename() on the result to obtain the filename itself.

Details

The default dir (".") indicates the working directory.

If ignore_case is FALSE and no case-sensitive match is found, the error message indicates if any case-insensitive match is present.

In contrast to the default of list.files(), get_file_path() also finds 'hidden' files, i.e., files with names that start with a dot.

Paths will be normalized to ensure they still work if the working directory changes. "/" instead of "\\" is used as argument winslash such that the returned path can be used in Windows' file system.

See also

Examples

# Create files in a temporary directory so we know what is present.
my_tempfiles <- tempfile(pattern = c("some_filename", "another_filename"),
                         fileext = ".txt")
# Create the files
file.create(my_tempfiles)
#> [1] TRUE TRUE

get_file_path(dir = tempdir(), pattern = "some_file")
#> Using file '/tmp/RtmpETjhnx/some_filename19bb7f4ff4ec.txt'
#> [1] "/tmp/RtmpETjhnx/some_filename19bb7f4ff4ec.txt"

# The same file is found if case-insensitive matching is used:
get_file_path(dir = tempdir(), pattern = "SOME_FILE", ignore_case = TRUE)
#> Using file '/tmp/RtmpETjhnx/some_filename19bb7f4ff4ec.txt'
#> [1] "/tmp/RtmpETjhnx/some_filename19bb7f4ff4ec.txt"

# Error reporting the presence of a case-insensitive match.
try(get_file_path(dir = tempdir(), pattern = "SOME_FILE", ignore_case = FALSE))
#> Error in get_file_path(dir = tempdir(), pattern = "SOME_FILE", ignore_case = FALSE) : 
#>   No case-sensitive matches to pattern 'SOME_FILE' are present in directory
#> '/tmp/RtmpETjhnx'.
#> However, a case-insensitive match to 'pattern' is present: 'some_filename19bb7f4ff4ec.txt'.

# Error reporting no match found.
try(get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",
                 ignore_case = TRUE))
#> Error in get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",  : 
#>   No matches to pattern 'missing_filename_abcde' are present in directory
#> '/tmp/RtmpETjhnx'.
try(get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",
                 ignore_case = FALSE))
#> Error in get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",  : 
#>   No case-sensitive matches to pattern 'missing_filename_abcde' are present in directory
#> '/tmp/RtmpETjhnx'.
#> No case-insensitive match is present either.

# Error if multiple matches are present.
try(get_file_path(dir = tempdir(), pattern = "_filename"))
#> Error in get_file_path(dir = tempdir(), pattern = "_filename") : 
#>   Multiple matches to pattern '_filename' are present in directory
#> '/tmp/RtmpETjhnx': 'another_filename19bb64a0df44.txt', 'some_filename19bb7f4ff4ec.txt'!

# Clean up
unlink(x = my_tempfiles)
rm(my_tempfiles)