Check that x is a valid filename

Check that x is a filename that includes an extension and is likely to be a valid filename.

Usage

is_filename(filename)

Arguments

filename

character string with the filename, including the file extension.

Value

TRUE: an error occurs if filename is not a valid filename.

Details

is_filename() puts some restrictions on filenames so it can be used to check for valid filenames before creating a file:

filename should include a file extension. The compression extensions ".gz", ".bz2" or ".xz" are not considered file extensions but are allowed to be present.

filename should not contain characters ", *, /, :, ?, \, |, <, >, nor any of the control characters (ASCII octal codes 000 through 037 and 177, see help("regex")).

filename should not start with a space.

The part of filename before the extension should not end with a space nor with a period.

These restrictions consider characters that would lead to an error in Windows because they are not allowed; characters that would lead to a mismatch between the created directory and the returned path because they are silently removed in Windows; and words that are reserved names in Windows.

In contrast to functions from checkinput, is_filename will produce an error if filename is not a valid filename.

Programming notes

Although not enforced by is_filename(), it is good practice to also avoid the characters +, ,, ;, =, [, ], !, $, #, @

Also avoid the characters {, }, (, ), ', %, &, , ^, ~.

Ways to make is_filename() even stricter:

do not allow filenames to start with a hyphen or underline

do not allow filenames to end with a hyphen or underline

case-insensitive matching to filename to determine if it exists? filename should not point to a directory (see utils::file_test(), get_filename.R, create_tempdir.R, create_path.R).

Impose a limit on the length: https://blog.r-project.org/2023/03/07/path-length-limit-on-windows/

See also function progutils::create_path(), usethis:::make_filename() and fs::path_real().

References

https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file

https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits

See also

Examples

is_filename("abcd.txt")
#> [1] TRUE
is_filename("abcd.txt.gz")
#> [1] TRUE

try(is_filename("abcd"))
#> Error in is_filename("abcd") : Empty filename or missing extension:
#> abcd
try(is_filename("abcd.gz"))
#> Error in is_filename("abcd.gz") : Empty filename or missing extension:
#> abcd.gz
try(is_filename("ab|cd.txt"))
#> Error in is_filename("ab|cd.txt") : 
#>   'filename' should not contain '"', '*', '/', ':', '?', '', '|', '<' or '>':
#> ab|cd.txt