Create a temporary directory

Create a temporary directory that can safely be removed.

Usage

create_tempdir(subdir = "subdir")

Arguments

subdir

A character string with the name of the temporary subdirectory to be created inside tempdir(). subdir should be a valid path pointing to a directory that not yet exists, see Details.

Value

The normalized path to the created temporary directory, returned invisibly.

Details

subdir is created inside tempdir() and an error is thrown if it already exists. This ensures that programmatically removing the created directory later on does not remove files that are still needed by other processes (which could happen when removing the directory returned by tempdir(), for example because RStudio also uses that directory).

It is possible to create subdirectories inside a not-yet existing directory (e.g., to create <tempdir>/output/outputsub if <tempdir>/output does not yet exist.

Side effects

The requested temporary directory is created if does not yet exist. An error is thrown if the directory already exists or creating the directory failed.

See also

Examples

tempdir()
#> [1] "/tmp/RtmpdBl9ox"
# Create a directory inside the directory returned by 'tempdir()'
(tempdir_std <- create_tempdir(subdir = "examplesubtempdir"))
#> [1] "/tmp/RtmpdBl9ox/examplesubtempdir"

# Error if the directory already exists
try(create_tempdir(subdir = "examplesubtempdir"))
#> Error in create_tempdir(subdir = "examplesubtempdir") : 
#>   Temporary directory already exists: change 'subdir' ('examplesubtempdir'): /tmp/RtmpdBl9ox/examplesubtempdir

# It is possible to create recursive directories
(tempdir_recursive <- create_tempdir(subdir = file.path("abc", "def")))
#> [1] "/tmp/RtmpdBl9ox/abc/def"

# Clean up
unlink(c(tempdir_std, dirname(tempdir_recursive)), recursive = TRUE)
rm(tempdir_recursive, tempdir_std)