Create a character vector from a string with quotation marks

Create a character vector from a string with quotation marks

Usage

unpaste_unquote(x, collapse = c(", ", "; "), quotemarks = c("'", "\""))

Arguments

x

A character string

collapse

Character vector with elements that were used to collapse values when x was created, and are now used to split x on. Can be character(0) to leave x as a character string.

quotemarks

Character vector with quotation marks to be removed from x. Can be character(0) to not remove any quotation marks.

Value

x without the quotation marks and split into a vector.

Details

unpaste_unquote() is not the exact reverse of paste_quoted(): it does not restore zero-length elements to their original value, e.g., "'NULL'" to NULL, or "'character(0)'" to character(0).

See also

Examples

x <- paste_quoted(c("ff", "gG", "HH"))
x
#> [1] "'ff', 'gG', 'HH'"
unpaste_unquote(x = x, collapse = ", ", quotemarks = "'")
#> [1] "ff" "gG" "HH"
unpaste_unquote(x = x, collapse = ", ", quotemarks = "\"")
#> [1] "'ff'" "'gG'" "'HH'"
unpaste_unquote(x = x, collapse = character(0), quotemarks = c("'", "\""))
#> [1] "ff, gG, HH"
unpaste_unquote(x = x, collapse = character(0), quotemarks = character(0))
#> [1] "'ff', 'gG', 'HH'"